+++ module/cards/torgeternityPlayerHand.js
@@ -28,7 +28,6 @@ export default class torgeternityPlayerHand extends foundry.applications.sheets.
   static PARTS = {
     normal: { template: 'systems/torgeternity/templates/cards/torgeternityPlayerHand.hbs', scrollable: ["ol[data-cards]"] },
     lifelike: { template: "systems/torgeternity/templates/cards/torgeternityPlayerHand_lifelike.hbs" },
-    footer: { template: "templates/generic/form-footer.hbs" }
   }
 
   /**
@@ -41,7 +40,16 @@ export default class torgeternityPlayerHand extends foundry.applications.sheets.
     for (const card of context?.document.cards) {
       card.typeLoc = game.i18n.localize(`torgeternity.cardTypes.${card.type}`);
     }
+    context.buttons = this._prepareButtons();
     return context;
+  }
+
+  _configureRenderParts(options) {
+    const part = this.document.flags.torgeternity.lifelike ? 'lifelike' : 'normal';
+    let parts = {};
+    parts[part] = foundry.utils.deepClone(this.constructor.PARTS[part]);
+    parts[part].root = true;
+    return parts;
   }
 
   async _preparePartContext(partId, context, options) {
@@ -309,9 +317,13 @@ export default class torgeternityPlayerHand extends foundry.applications.sheets.
    *
    */
   async drawCosmDialog() {
-    const data = {};
-    data.decks = game.settings.get('torgeternity', 'deckSetting');
-    data.unused = torgeternityDeck.UNUSED_DECK_ID;
+    const data = { decks: {} };
+    const decks = foundry.utils.duplicate(game.settings.get('torgeternity', 'deckSetting'));
+    for (const key in decks) {
+      if (CONFIG.torgeternity.cosmDecks[key] && decks[key] !== torgeternityDeck.UNUSED_DECK_ID) {
+        data.decks[decks[key]] = `torgeternity.cosmDecks.${key}`;
+      }
+    }
     const html = await foundry.applications.handlebars.renderTemplate(
       'systems/torgeternity/templates/cards/drawCosmDialog.hbs',
       data